+++ packages/components/src/components/TunnelEntry/TunnelEntry.tsx
@@ -1,14 +1,14 @@
-import { TunnelEntry as ReactTunnelEntry } from "@mittwald/react-tunnel";
-import type { TunnelEntryProps as ReactTunnelEntryProps } from "@mittwald/react-tunnel/types";
 import type { FC } from "react";
+import { TunnelEntry as ReactTunnelEntry } from "@mittwald/react-tunnel";
+import type { TunnelEntryProps } from "@mittwald/react-tunnel";
 
-export type TunnelEntryProps = ReactTunnelEntryProps;
+export type { TunnelEntryProps } from "@mittwald/react-tunnel";
 
 /**
  * @flr-generate all
  * @flr-clear-props-context
  */
-export const TunnelEntry: FC<ReactTunnelEntryProps> = (props) => (
+export const TunnelEntry: FC<TunnelEntryProps> = (props) => (
   <ReactTunnelEntry {...props} />
 );
