+++ packages/react-tunnel/package.json
@@ -16,10 +16,6 @@
     ".": {
       "types": "./dist/types/index.d.ts",
       "import": "./dist/index.js"
-    },
-    "./types": {
-      "types": "./dist/types/types.d.ts",
-      "import": "./dist/types.js"
     }
   },
   "types": "./dist/types/index.d.ts",
